requirements.txt
Django==1.6.1
Jinja2==2.7.2
Landscape-Client==14.01
MarkupSafe==0.18
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
South==1.0
Twisted-Core==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
beautifulsoup4==4.2.1
chardet==2.0.1
colorama==0.2.5
configobj==4.7.2
cvxopt==1.1.4
decorator==3.4.0
django-tables2==0.15.0
gevent==1.0
greenlet==0.4.2
gunicorn==17.5
gyp==0.1
html5lib==0.999
ipython==1.2.1
joblib==0.7.1
lxml==2.3.2
matplotlib==1.3.1
nose==1.3.1
numexpr==2.2.2
numpy==1.9.1
openpyxl==1.7.0
pandas==0.13.1
patsy==0.2.1
pexpect==3.1
psycopg2==2.4.5
pyOpenSSL==0.13
pycurl==7.19.3
pyparsing==2.0.1
pyserial==2.6
python-apt==0.9.3.5
python-dateutil==1.5
python-debian==0.1.21-nmu2ubuntu2
python-docx==0.7.4
pytz==2012c
pyzmq==14.0.1
requests==2.2.1
scipy==0.13.3
simplegeneric==0.8.1
simplejson==3.3.1
six==1.5.2
ssh-import-id==3.21
statsmodels==0.5.0
sympy==0.7.4.1
tables==3.1.1
tornado==3.1.1
urllib3==1.7.1
virtualenv==1.11.4
wsgiref==0.1.2
xlrd==0.9.2
xlwt==0.7.5
zope.interface==4.0.5
